% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{build_scenario}
\alias{build_scenario}
\title{Experiment method for building future scenarios for a model}
\usage{
build_scenario(
  e,
  base_t,
  finish_t,
  na.constant,
  inputs,
  stock_pars,
  initial_values
)
}
\arguments{
\item{e}{A rhapsody experiment object}

\item{base_t}{An integer denoting the starting timestep for the scenario}

\item{finish_t}{An integer denoting the finishing timestep for the scenario}

\item{na.constant}{Logical: should inputs and stock parameters not included in scenario input files be held constant?}

\item{inputs}{A dataframe: each column should be named after a model input}

\item{stock_pars}{A list of dataframes: each dataframe in the list should be named after an internal stock parameter, with the column names corresponding to the relevant model stock}

\item{initial_values}{A dataframe to initialise stock parameters}
}
\value{
A rhapsody experiment object
}
\description{
Experiment method for building future scenarios for a model
}
\examples{
inputs     <- data.frame(Year = 2020:2050, i_soybean_demand = seq(1, 2, by = 1/30))
stock_pars <- list(s_yield = data.frame(Year = 2020:2050, soybeans = seq(1, 1.25, by = 0.25/30)))
sd.exp     <- build_scenario(sd.exp, base_t = 2020, finish_t = 2050, na.constant = T, inputs = inputs, stock_pars=  stock_pars, initial_values= start_vals)
}
