% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{Experiment}
\alias{Experiment}
\title{Experiment class: scenario building, batch running & parameter selection}
\usage{
Experiment(.model, .parameter_set, .initial, .n_runs, .n_cores = 1)
}
\arguments{
\item{.model}{A rhapsody model object}

\item{.parameter_set}{A data frame with 1 column per model parameter}

\item{.initial}{A data frame to set initial stock parameters, similar to that in the init_model function}

\item{.n_runs}{Integer, number of model runs to conduct from parameter set}

\item{.n_cores}{Integer, used to set cores for parallel processing}
}
\value{
A rhapsody experiment object
}
\description{
Experiment class: scenario building, batch running & parameter selection
}
\examples{

Priors <- data.frame(p_land_trans = runif(1000, 0.00001, 0.001), p_price_update = runif(1000, 0.01, 0.1))
sd.exp <- Experiment(sd.mod, Priors, start_vals, nrow(Priors))
}
