% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{class}
\name{flow-class}
\alias{flow-class}
\alias{flow}
\title{Initialise a model stock}
\arguments{
\item{f_name}{A character; if using standard taxonomy starting with 'f_'}

\item{f_id}{A character giving a unique object identifier}

\item{f_from_stocks}{A character naming stocks and inputs that are upstream in the model structure}

\item{f_to_stocks}{A character naming stocks that are downstream in the model structure}

\item{f_function}{A character naming the function to be used to run the flow; this can be loaded from a package or defined in the global environment}

\item{f_values}{An empty list to be filled by flow function inputs and outputs}
}
\value{
A rhapsody flow
}
\description{
Initialise a model stock
}
\examples{
f_deforest <- flow(f_name = 'f_deforest', f_id = 'f1', f_from_stocks = c('s_soybean, s_forest'),
             f_to_stocks = c('s_soybean, s_forest'), f_function = 'deforest')
}
