% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_system.R
\name{flows_from_key}
\alias{flows_from_key}
\title{Create a list of flows from a flow key}
\usage{
flows_from_key(key, parse_stocks = T)
}
\arguments{
\item{key}{A dataframe generated either in the global environment or through the template csv file}

\item{parse_stocks}{A logical: should the stocks to and stocks from columns be parsed into a character vector}
}
\value{
A list of flow objects
}
\description{
Create a list of flows from a flow key
}
\examples{
flows_from_key(flow_key, parse_stocks = T)
}
