% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{get_flow_args}
\alias{get_flow_args}
\title{get values of flow arguments from model object}
\usage{
get_flow_args(x, mod, parse_args)
}
\arguments{
\item{x}{A rhapsody flow object}

\item{mod}{A rhapsody model object}

\item{parse_args}{A logical; should only the data elements of relevant objects be returned?}
}
\value{
A rhapsody flow
}
\description{
get values of flow arguments from model object
}
\examples{
f_deforest        <- get_flow_args(f_deforest, sd.mod, parse_args = F)
sd.mod@flows[[1]] <- get_flow_args(sd.mod@flows[[1]], sd.mod)
}
