% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{init_model}
\alias{init_model}
\title{Initiliase a model - the core class for running a rhapsody system dynamics model. For a more detailed explanation see 'model' vignette}
\usage{
init_model(
  stock_list,
  flows_list,
  input_list,
  output_list,
  parameter_list,
  initial_stock_vals,
  run = list(nsteps = 1, step = 1),
  taxon = list(s_ = "stocks", f_ = "flows", ff = "family_flows", p_ = "parameters", i_ =
    "inputs", o_ = "outputs"),
  load_stock_dat = T
)
}
\arguments{
\item{stock_list}{A list of stocks; typically the result of stocks_from_key}

\item{input_list}{A list of inputs; typically the result of inputs_from_key}

\item{output_list}{A list of outputs; typically the result of outputs_from_key}

\item{parameter_list}{A named list of model parameters, each beginning with the 'p_' taxonomic identifier}

\item{initial_stock_vals}{A three column dataframe; the 1st column specifies the relevant stock, the second the relevant stock parameter, and the third the initialising value}

\item{run}{a named list containing two integers: nsteps and step. These specify the number of model timesteps and the name -e.g. the year or month - of the 1st model timestep}

\item{taxon}{a named list used to specify model naming conventions; it is recommended to leave these as defaults.}

\item{load_stock_dat}{a logical - should model stock parameter data be loaded on initialisation?}

\item{flow_list}{A list of flows; typically the result of flows_from_key}
}
\value{
A rhapsody model object
}
\description{
Initiliase a model - the core class for running a rhapsody system dynamics model. For a more detailed explanation see 'model' vignette
}
\examples{
mod    <- init_model(stocks_from_key(s_key), flows_from_key(f_key),
           inputs_from_key(i_key), outputs_from_key(o_key),
             initial_stock_vals = start_vals,
              list(p_land_conversion = 0.05, p_price_update = 0.01),
               run = list(nsteps = 31, step = 1990))

}
