% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stock.R
\name{init_pars}
\alias{init_pars}
\title{Load stock parameters: these are those that are dynamically updated by other model objects}
\usage{
init_pars(x, isv)
}
\arguments{
\item{x}{A rhapsody stock}

\item{isv}{A dataframe, typically that passed to the initial_parameters argument of init_model}
}
\value{
A rhapsody stock object
}
\description{
Load stock parameters: these are those that are dynamically updated by other model objects
}
\examples{
s_soybeans <- init_pars(s_soybeans, start_vals)
}
