% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stock.R
\name{init_stock}
\alias{init_stock}
\title{Initialise a model stock}
\usage{
init_stock(
  name,
  id,
  family,
  flows_out,
  flows_in,
  internal_pars,
  parameters = list()
)
}
\arguments{
\item{name}{A character string; if using standard taxonomy starting with 's_'}

\item{id}{A character string giving a unique object identifier}

\item{family}{A character string assigning the stock to a family of stocks}

\item{flows_out}{A character naming stocks and outputs that are downstream in the model structure}

\item{flows_in}{A character naming stocks and inputs that are uptream in the model structure}

\item{internal_pars}{A character vector naming the stocks internal parameters}

\item{parameters}{An empty list to be filled by model intial parameters}
}
\value{
A rhapsody stock object
}
\description{
Initialise a model stock
}
\examples{
s_soybeans <- init_stock(s_name = 's_soybean',
    s_family = 'cropland', s_id = 's1',
         s_flow_out = 's_soybean', s_flow_in = 'i_soybean_demand',
           s_parameters = c('area', 'yield', 'price'), s_internal_pars = '')
}
