% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_system.R
\name{inputs_from_key}
\alias{inputs_from_key}
\title{Create a list of inputs from an input key}
\usage{
inputs_from_key(key, path = NA, data_type = "generic")
}
\arguments{
\item{key}{A dataframe generated either in the global environment or through the template csv file}

\item{path}{A character string giving the location of input data files for the model, or NA for the current working directory}

\item{data_type}{A character string naming the data format of the input file; currently on generic supported}
}
\value{
A list of input objects
}
\description{
Create a list of inputs from an input key
}
\examples{
inputs_from_key(input_key, path = 'my/file/path')
}
