% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stock.R
\name{load_stock_data}
\alias{load_stock_data}
\title{Load stock internal parameter data set}
\usage{
load_stock_data(x, path)
}
\arguments{
\item{x}{A rhapsody stock}

\item{path}{Either NA, which uses defaults to 'stockname.csv', or a file path giving the location of the stock internal data}
}
\value{
A rhapsody stock object
}
\description{
Load stock internal parameter data set
}
\examples{
s_soybeans <- load_stock_data(s_soybeans, NA)
}
