% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_system.R
\name{make_system}
\alias{make_system}
\title{A utility function for checking the system as specified. Not used in simulation modelling.}
\usage{
make_system(stocks_list, flows_list, tidy = F)
}
\arguments{
\item{stocks_list}{A list of stocks; typically the output of stocks_from_key}

\item{flows_list}{A list of flows; typically the output of flows_from_key}

\item{tidy}{A logical; passed to igraph for presentation. Shortens stock names.}
}
\value{
An igraph::graph object
}
\description{
A utility function for checking the system as specified. Not used in simulation modelling.
}
\examples{
plot(make_system(s_list, f_list, tidy = T))
}
