% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_system.R
\name{outputs_from_key}
\alias{outputs_from_key}
\title{Create a list of outputs from an output key}
\usage{
outputs_from_key(key)
}
\arguments{
\item{key}{A dataframe generated either in the global environment or through the template csv file}
}
\value{
A list of output objects
}
\description{
Create a list of outputs from an output key
}
\examples{
outputs_from_key(output_key)
}
