% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{reset_time}
\alias{reset_time}
\title{Helper for resetting the model timestep to 1}
\usage{
reset_time(x)
}
\arguments{
\item{x}{A rhapsody model object}
}
\value{
A rhapsody model object
}
\description{
Helper for resetting the model timestep to 1
}
\examples{
sd.mod    <- run_mod(sd.mod)
sd.mod$parameters$p_land_trans <- 0.01 ### some arbitrary change
sd.mod    <- reset_time(sd.mod)
sd.mod    <- run_mod(sd.mod)
}
