% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Function for running a model experiment a single time. Useful for checking outputs before a large batch-run}
\usage{
run_experiment(e)
}
\arguments{
\item{e}{A rhapsody experiment object}
}
\value{
A rhapsody experiment object, with results stored in the results slot
}
\description{
Function for running a model experiment a single time. Useful for checking outputs before a large batch-run
}
\examples{

sd.exp <- run_experiment(sd.exp)
}
