% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{run_flow}
\alias{run_flow}
\title{runs flow}
\usage{
run_flow(x, ...)
}
\arguments{
\item{x}{A rhapsody flow}
}
\value{
A rhapsody flow
}
\description{
runs flow
}
\examples{
f_deforest <- run_flow(f_deforest)
}
