% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{run_mod}
\alias{run_mod}
\title{run_mod executes the scheduler using the information in the runs argument passed to a rhapsody model}
\usage{
run_mod(x)
}
\arguments{
\item{x}{A rhapsody model object}
}
\value{
A rhapsody model object
}
\description{
run_mod executes the scheduler using the information in the runs argument passed to a rhapsody model
}
\examples{
sd.mod    <- run_mod(sd.mod)
}
