% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{scheduler}
\alias{scheduler}
\title{The scheduler updates the inputs and stock internal parameters; it then runs the model flows and captures model outputs}
\usage{
scheduler(x, ...)
}
\arguments{
\item{x}{A rhapsody model object}
}
\value{
A rhapsody model object
}
\description{
The scheduler updates the inputs and stock internal parameters; it then runs the model flows and captures model outputs
}
\examples{
sd.mod    <- scheduler(sd.mod)
}
