% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{select_pars}
\alias{select_pars}
\title{Select model parameters through history matching}
\usage{
select_pars(e, data, method = "best", ...)
}
\arguments{
\item{e}{A rhapsody experiment object}

\item{data}{A dataframe containing observations}

\item{method}{A character string; either 'best' or 'tolerance'. Currently only best is implemented}

\item{.subset}{A character string specifying which model outputs to history match}

\item{.frac}{A double (0-1); determines what fraction of model parameters to select if 'best' is chosen method}

\item{.scale}{A logical: should observations & model outputs be scaled before rmse is calculated?}
}
\value{
A named list with two parts: $runs, which gives the row number of the selected runs; and $metric, which reports the model selection metric for the chosen runs
}
\description{
Select model parameters through history matching
}
\examples{

### fake data for history matching
dat              <- data.frame(soybean_area = seq(1000, 500000, by = (500000-1000)/30), forest_area = seq(1000000, 500000, by = 0-(500000/30)))
noise            <- sapply(seq(1000, 500000, by = (500000-1000)/30), function(z) {rnorm(1, z, z/3)})
dat$soybean_area <- dat$soybean_area + noise
dat$forest_area  <- dat$forest_area - noise

### filter the parameters based on their fit to the observations
result <- select_pars(sd.exp, data = dat, method = 'best', .frac = 0.01, .scale = T,
                     .subset = sapply(sd.exp@mod@outputs, function(z) {z@o_name})[c(1:2)])

### look at the results
summary(Priors)
summary(Priors[result$runs, ])
}
