% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{set_flow_output}
\alias{set_flow_output}
\title{runs flow}
\usage{
set_flow_output(x, mod)
}
\arguments{
\item{x}{A rhapsody flow}

\item{mod}{A rhapsody model}
}
\value{
A rhapsody model object
}
\description{
runs flow
}
\examples{
sd.mod <- set_flow_output(sd.mod@flows[[1]], sd.mod)
}
