% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{single_experiment}
\alias{single_experiment}
\title{Function for running a model experiment a single time. Useful for checking outputs before a large batch-run}
\usage{
single_experiment(e, run_numb)
}
\arguments{
\item{e}{A rhapsody experiment object}

\item{run_numb}{Integer - used to pick a parameter set from within the declared ranges}
}
\value{
A rhapsody experiment object
}
\description{
Function for running a model experiment a single time. Useful for checking outputs before a large batch-run
}
\examples{

sd.exp <- single_experiment(sd.exp, 128)
}
