% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{spin_up}
\alias{spin_up}
\title{Runs a model multiple times with its first set of inputs; used to find convergence / equilibrium model state}
\usage{
spin_up(x, n)
}
\arguments{
\item{x}{A rhapsody model object}

\item{n}{An integer}
}
\value{
A rhapsody model object
}
\description{
Runs a model multiple times with its first set of inputs; used to find convergence / equilibrium model state
}
\examples{
sd.mod    <- spin_up(sd.mod)
}
