% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_system.R
\name{stocks_from_key}
\alias{stocks_from_key}
\title{Create a list of stocks from a stock key}
\usage{
stocks_from_key(key, parse_flows = T)
}
\arguments{
\item{key}{A dataframe generated either in the global environment or through the template csv file}

\item{parse_flows}{A logical: should the flows to and flows from columns be parsed into a character vector}
}
\value{
A list of stock objects
}
\description{
Create a list of stocks from a stock key
}
\examples{
stocks_from_key(stock_key, parse_flows = T)
}
