% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stock.R
\name{update_internal_par}
\alias{update_internal_par}
\title{Load stock parameters: these are those that are dynamically updated by other model objects}
\usage{
update_internal_par(x, mod)
}
\arguments{
\item{x}{A rhapsody stock object}

\item{mod}{A rhapsody model object}
}
\value{
A rhapsody stock object
}
\description{
Load stock parameters: these are those that are dynamically updated by other model objects
}
\examples{
s_soybeans <- update_internal_par(s_soybeans, sd.mod)
}
